+++ examples/sapper-rollup/src/routes/about.svelte
@@ -1,8 +1,0 @@
-<template lang="pug">
-  svelte:head
-    title About
-
-  h1 About this site
-
-  p This is the 'about' page. There's not much here.
-</template>