+++ examples/svelte-rollup/rollup.config.ts
@@ -1,7 +1,0 @@
-
-import sveltePreprocess from 'svelte-preprocess';
-// const sveltePreprocess = require('../../dist/index.js')
-
-sveltePreprocess({
-  replace: [['foo', 'bar']]
-})