+++ src/transformers/babel.ts
@@ -20,6 +20,17 @@ const transformer: Transformer<Options.Babel> = async ({
     minified: false,
     ast: false,
     code: true,
+    caller: {
+      name: 'svelte-preprocess',
+      supportsStaticESM: true,
+      supportsDynamicImport: true,
+      // this isn't supported by Svelte but let it error with a good error on this syntax untouched
+      supportsTopLevelAwait: true,
+      // todo: this can be enabled once all "peer deps" understand this
+      // this syntax is supported since rollup@1.26.0 and webpack@5.0.0-beta.21
+      // supportsExportNamespaceFrom: true,
+      ...options?.caller,
+    },
   } as TransformOptions;
 
   const result = await transformAsync(content, babelOptions);