+++ test/autoProcess/externalFiles.test.ts
@@ -51,12 +51,12 @@ describe('external files', () => {
       }),
     ];
 
-    expect(markup.code).toContain(getFixtureContent('template.html'));
-    expect(script.code).toContain(getFixtureContent('script.js'));
-    expect(style.code).toContain(getFixtureContent('style.css'));
-    expect(markup.dependencies).toContain(getFixturePath('template.html'));
-    expect(script.dependencies).toContain(getFixturePath('script.js'));
-    expect(style.dependencies).toContain(getFixturePath('style.css'));
+    expect(markup?.code).toContain(getFixtureContent('template.html'));
+    expect(script?.code).toContain(getFixtureContent('script.js'));
+    expect(style?.code).toContain(getFixtureContent('style.css'));
+    expect(markup?.dependencies).toContain(getFixturePath('template.html'));
+    expect(script?.dependencies).toContain(getFixturePath('script.js'));
+    expect(style?.dependencies).toContain(getFixturePath('style.css'));
   });
 
   it('should support self-closing tags', async () => {
@@ -65,26 +65,43 @@ describe('external files', () => {
       filename: resolve(__dirname, '..', 'App.svelte'),
     });
 
-    expect(markup.code).toContain(getFixtureContent('template.html'));
+    expect(markup?.code).toContain(getFixtureContent('template.html'));
   });
 
-  REMOTE_JS.forEach((url) => {
-    it(`should not attempt to locally resolve ${url}`, async () => {
-      const input = `<div></div><script src="${url}"></script>`;
-
-      const preprocessed = await preprocess(input, sveltePreprocess());
-
-      expect(preprocessed.toString?.()).toContain(input);
-      expect(preprocessed.dependencies).toHaveLength(0);
-    });
-  });
-
-  it("should warn if local file don't exist", async () => {
+  it("warns if local file don't exist", async () => {
     const input = `<style src="./missing-potato"></style>`;
 
     await preprocess(input, sveltePreprocess());
 
     expect(warnSpy).toHaveBeenCalledWith(
+      expect.stringContaining('was not found'),
+    );
+  });
+
+  REMOTE_JS.forEach((url) => {
+    it(`ignores remote path "${url}"`, async () => {
+      const input = `<div></div><script src="${url}"></script>`;
+
+      const preprocessed = await preprocess(input, sveltePreprocess());
+
+      expect(preprocessed.toString?.()).toContain(input);
+      expect(preprocessed?.dependencies).toHaveLength(0);
+      expect(warnSpy).not.toHaveBeenCalledWith(
+        expect.stringContaining('was not found'),
+      );
+    });
+  });
+
+  it('ignores external source if path is not relative', async () => {
+    const input = `<style src="/root-potato"></style>`;
+
+    await preprocess(input, sveltePreprocess());
+
+    const preprocessed = await preprocess(input, sveltePreprocess());
+
+    expect(preprocessed.toString?.()).toContain(input);
+    expect(preprocessed?.dependencies).toHaveLength(0);
+    expect(warnSpy).not.toHaveBeenCalledWith(
       expect.stringContaining('was not found'),
     );
   });