+++ test/transformers/babel.test.ts
@@ -32,12 +32,39 @@ $: bar = foo?.b ?? 120
 
     const preprocessed = await preprocess(template, opts);
 
-    expect(preprocessed.code).toMatchInlineSnapshot(`
-      "<script>var _foo$b;
+    expect(preprocessed.code).toBe(`<script>var _foo$b;
+var foo = {};
+$: bar = (_foo$b = foo == null ? void 0 : foo.b) != null ? _foo$b : 120;</script>`);
+  });
 
-      var foo = {};
+  it('should not transpile import/export syntax with preset-env', async () => {
+    const template = `<script>
+import foo from './foo'
+$: bar = foo?.b ?? 120
+</script>`;
 
-      $: bar = (_foo$b = foo == null ? void 0 : foo.b) != null ? _foo$b : 120;</script>"
-    `);
+    const opts = sveltePreprocess({
+      babel: {
+        presets: [
+          [
+            '@babel/preset-env',
+            {
+              loose: true,
+              targets: {
+                esmodules: true,
+              },
+            },
+          ],
+        ],
+      },
+    });
+
+    const preprocessed = await preprocess(template, opts);
+
+    expect(preprocessed.code).toBe(
+      `<script>var _foo$b;
+import foo from './foo';
+$: bar = (_foo$b = foo == null ? void 0 : foo.b) != null ? _foo$b : 120;</script>`,
+    );
   });
 });